Moved sample_avg_dih from zmat/base/extra to zmat/_ring to solve circular import issue
